Replaced storing groceries lists with integration of Real-time database Firebase.
Instead of sending the data from the form to the home screen through navigation, sending POST request to add the created data to the database.

Groceries are now fetched from the database and added to a local list which displays the list items. However, to avoid unecessary requests when going back to the home screen after creating an item, instead of sending a request to display the new item I sent the data including the id from Firebase (Taken from the response of the POST) through navigation.